+++ src/clj/lambdacd/steps/git.clj
@@ -94,7 +94,7 @@
     (temp-util/with-temp repo-location
       (if (zero? checkout-exit-code)
         (let [execute-steps-result (execution/execute-steps steps (assoc args :cwd repo-location) ctx
-                                                            :unify-results-fn (support/unify-only-status unify/successful-when-all-successful)
+                                                            :unify-results-fn (unify/unify-only-status unify/successful-when-all-successful)
                                                             :is-killed (:is-killed ctx))
               result-with-checkout-output (assoc execute-steps-result :out (:out checkout-result))
               step-ids-and-outputs (:outputs execute-steps-result)