+++ test/clj/lambdacd/stress_test.clj
@@ -5,13 +5,14 @@
             [lambdacd.steps.control-flow :refer [in-parallel]]
             [lambdacd.util.internal.temp :as temp-utils]
             [lambdacd.steps.support :as support]
-            [lambdacd.util :as util]))
+            [lambdacd.util :as util]
+            [lambdacd.stepsupport.output :as output]))
 
 
 (def bacon-ipsum "Bacon ipsum dolor amet meatloaf tongue flank biltong ground round. Sirloin venison sausage kielbasa andouille strip steak tri-tip. Meatloaf biltong capicola shoulder. Sausage shankle pancetta prosciutto porchetta short loin.\n\n")
 
 (defn print-1000-lines-in-clj [args ctx]
-  (support/capture-output ctx
+  (output/capture-output ctx
     (doseq [i (range 1 1000)]
       (do
         (println i ")" bacon-ipsum))))